histogram parameters moved in another file, test for dimension dropdown added and valid states for projections and min-max value pairs
